requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
et_xmlfile==2.0.0
greenlet==3.3.2
idna==3.11
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.47
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
